Añadir los productos al carrito de compras y mostralos en la vista carrito

+++ src/main/java/com/curso/ecommerce/model/DetalleOrden.java
@@ -12,7 +12,7 @@ public class DetalleOrden {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String nombre;
-    private Double cantidad;
+    private Integer cantidad;
     private Double precio;
     private Double total;
 
@@ -25,7 +25,7 @@ public class DetalleOrden {
     public DetalleOrden() {
     }
 
-    public DetalleOrden(Integer id, String nombre, Double cantidad, Double precio, Double total) {
+    public DetalleOrden(Integer id, String nombre, Integer cantidad, Double precio, Double total) {
         this.id = id;
         this.nombre = nombre;
         this.cantidad = cantidad;
@@ -49,11 +49,11 @@ public class DetalleOrden {
         this.nombre = nombre;
     }
 
-    public Double getCantidad() {
+    public Integer getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(Double cantidad) {
+    public void setCantidad(Integer cantidad) {
         this.cantidad = cantidad;
     }
